# Un-unroll ukernel C+intrinsics code.