@bjacob 2023-09-03
Benchmarks on AMD Zen4:

Before:

```
BM_mmt4d_f32f32f32_tile_8x8x1_avx2_fma/real_time           0.234 us        0.234 us      4194303 items_per_second=140.165G/s
BM_mmt4d_f32f32f32_tile_16x16x1_avx512_base/real_time      0.790 us        0.790 us      1048575 items_per_second=165.87G/s
BM_mmt4d_f16f16f32_tile_8x8x1_avx2_fma/real_time           0.401 us        0.401 us      2097151 items_per_second=81.6161G/s
BM_mmt4d_f16f16f32_tile_16x16x1_avx512_base/real_time       1.58 us         1.58 us       524287 items_per_second=82.7239G/s
BM_mmt4d_f16f16f16_tile_8x8x1_avx2_fma/real_time           0.407 us        0.407 us      2097151 items_per_second=80.5982G/s
BM_mmt4d_f16f16f16_tile_16x16x1_avx512_base/real_time       1.58 us         1.58 us       524287 items_per_second=82.7024G/s
BM_mmt4d_i8i8i32_tile_8x8x2_avx2_fma/real_time             0.354 us        0.354 us      2097151 items_per_second=185.207G/s
BM_mmt4d_i8i8i32_tile_16x16x2_avx512_base/real_time         1.12 us         1.12 us      1048575 items_per_second=233.833G/s
BM_mmt4d_i8i8i32_tile_16x16x2_avx512_vnni/real_time        0.799 us        0.799 us      1048575 items_per_second=328.26G/s
```

After:

```
BM_mmt4d_f32f32f32_tile_8x8x1_avx2_fma/real_time           0.237 us        0.237 us      4194303 items_per_second=138.541G/s
BM_mmt4d_f32f32f32_tile_16x16x1_avx512_base/real_time      0.794 us        0.794 us      1048575 items_per_second=165.059G/s
BM_mmt4d_f16f16f32_tile_8x8x1_avx2_fma/real_time           0.404 us        0.404 us      2097151 items_per_second=81.0356G/s
BM_mmt4d_f16f16f32_tile_16x16x1_avx512_base/real_time       1.92 us         1.92 us       524287 items_per_second=68.1921G/s
BM_mmt4d_f16f16f16_tile_8x8x1_avx2_fma/real_time           0.409 us        0.409 us      2097151 items_per_second=80.0502G/s
BM_mmt4d_f16f16f16_tile_16x16x1_avx512_base/real_time       1.92 us         1.92 us       524287 items_per_second=68.1674G/s
BM_mmt4d_i8i8i32_tile_8x8x2_avx2_fma/real_time             0.358 us        0.358 us      2097151 items_per_second=183.259G/s
BM_mmt4d_i8i8i32_tile_16x16x2_avx512_base/real_time         1.13 us         1.13 us      1048575 items_per_second=232.859G/s
BM_mmt4d_i8i8i32_tile_16x16x2_avx512_vnni/real_time        0.803 us        0.802 us      1048575 items_per_second=326.657G/s
```

So the only difference is the `f16f16` kernel (`f16f16f16` and `f16f16f32` share the same code), it's not a night-and-day difference (68 vs 82 Gflop/s), and it's a slow ukernel anyway (doing `f16` on an ISA that does not have `f16` arithmetic support, converting to `f32`) so people who really care about perf aren't on that path anyway.