@bjacob 2023-09-03
There is no difference on `arm64` using Clang.