@bjacob 2023-09-03
Re-unrolled that `f16f16fXX` code to recover the original performance so this PR is a performance no-op on Clang at least:

AMD Zen4 benchmark:

```
BM_mmt4d_f32f32f32_tile_8x8x1_avx2_fma/real_time           0.237 us        0.237 us      4194303 items_per_second=138.131G/s
BM_mmt4d_f32f32f32_tile_16x16x1_avx512_base/real_time      0.799 us        0.799 us      1048575 items_per_second=164.138G/s
BM_mmt4d_f16f16f32_tile_8x8x1_avx2_fma/real_time           0.408 us        0.408 us      2097151 items_per_second=80.398G/s
BM_mmt4d_f16f16f32_tile_16x16x1_avx512_base/real_time       1.60 us         1.60 us       524287 items_per_second=81.7837G/s
BM_mmt4d_f16f16f16_tile_8x8x1_avx2_fma/real_time           0.411 us        0.411 us      2097151 items_per_second=79.6776G/s
BM_mmt4d_f16f16f16_tile_16x16x1_avx512_base/real_time       1.60 us         1.60 us       524287 items_per_second=81.863G/s
BM_mmt4d_i8i8i32_tile_8x8x2_avx2_fma/real_time             0.357 us        0.357 us      2097151 items_per_second=183.387G/s
BM_mmt4d_i8i8i32_tile_16x16x2_avx512_base/real_time         1.13 us         1.13 us      1048575 items_per_second=231.928G/s
BM_mmt4d_i8i8i32_tile_16x16x2_avx512_vnni/real_time        0.806 us        0.806 us      1048575 items_per_second=325.31G/s
```